+++ extensions/omni_worker/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-09-13
+
+### Added
+
+* SQL autostart
+  functionality [#932](https://github.com/omnigres/omnigres/pull/932), [#947](https://github.com/omnigres/omnigres/pull/947)
+* Timer handler for worker tasks [#943](https://github.com/omnigres/omnigres/pull/943)
+
+### Fixed
+
+* SQL execution role restriction [#945](https://github.com/omnigres/omnigres/pull/945)
+* SQL handler should return after committing [#946](https://github.com/omnigres/omnigres/pull/946)
+
 ## [0.1.4] - 2025-09-04
 
 ### Fixed
@@ -33,7 +46,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Initial release
 
-[Unreleased]: https://github.com/omnigres/omnigres/commits/next/omni_sqlite
+[Unreleased]: https://github.com/omnigres/omnigres/commits/next/omni_worker
 
 [0.1.0]: [https://github.com/omnigres/omnigres/pull/856]
 
@@ -44,3 +57,5 @@ Initial release
 [0.1.3]: [https://github.com/omnigres/omnigres/pull/915]
 
 [0.1.4]: [https://github.com/omnigres/omnigres/pull/938]
+
+[0.2.0]: [https://github.com/omnigres/omnigres/pull/865]